moved auth code outside of server.ts
